Initial Commit
Added / edited basic configuration files to introduce the new airframe
Edited the board configuration for IMU, Compass and Serial Interfaces to work
Todo:
- GPS
- Baro
- PWM
- Camera(?)
- Tuning
- Fly

+++ Tools/setup/arch.sh
@@ -103,9 +103,20 @@ if [[ $INSTALL_NUTTX == "true" ]]; then
 		echo "arm-none-eabi-gcc-${NUTTX_GCC_VERSION} found, skipping installation"
 
 	else
-		echo "Installing arm-none-eabi-gcc-${NUTTX_GCC_VERSION}";
-		wget -O /tmp/gcc-arm-none-eabi-${NUTTX_GCC_VERSION}-linux.tar.bz2 https://armkeil.blob.core.windows.net/developer/Files/downloads/gnu-rm/${NUTTX_GCC_VERSION_SHORT}/gcc-arm-none-eabi-${NUTTX_GCC_VERSION}-x86_64-linux.tar.bz2 && \
+		GCC_FILENAME=/tmp/gcc-arm-none-eabi-${NUTTX_GCC_VERSION}-linux.tar.bz2
+		if test -f "$GCC_FILENAME"; then
+			echo "$GCC_FILENAME found, skipping download"
+		else
+			echo "Installing arm-none-eabi-gcc-${NUTTX_GCC_VERSION}";
+			wget -O /tmp/gcc-arm-none-eabi-${NUTTX_GCC_VERSION}-linux.tar.bz2 https://armkeil.blob.core.windows.net/developer/Files/downloads/gnu-rm/${NUTTX_GCC_VERSION_SHORT}/gcc-arm-none-eabi-${NUTTX_GCC_VERSION}-x86_64-linux.tar.bz2 && \
+		fi
+
+		GCC_INSTALLATION=/opt/gcc-arm-none-eabi-${NUTTX_GCC_VERSION}
+		if test -d "$GCC_INSTALLATION"; then
+			echo "$GCC_INSTALLATION found, skipping installation"
+		else
 			sudo tar -jxf /tmp/gcc-arm-none-eabi-${NUTTX_GCC_VERSION}-linux.tar.bz2 -C /opt/;
+		fi
 
 		# add arm-none-eabi-gcc to user's PATH
 		exportline="export PATH=/opt/gcc-arm-none-eabi-${NUTTX_GCC_VERSION}/bin:\$PATH"
